Update Intro to the Run in Postman button
I renamed the branch from atom, which was reflected in the GitHub app, but not the repo. As a result not all of the commints were in the right place. So, I've just deleted then put back everything I wrote to get a complete version of the final doc. Here's to hoping :/