Improving scannability/content UX by marrying sections
The page looks janky with all the h2s that have only a small amount of content and are then followed closely by an h3 or another h2. I want to regroup and order the info hierarchy so that it's scannable but also aethetically acceptable (not pleasing, but not off-putting at first glance). The two previous sections seemd like they could be easily married and logically connected via the first sentence of the last paragraph.