Deleted "Security considerations"
All content was rewritten in "The Run in Postman button"; No other pages linked to this page (tho because the page_id is so vague, git grep pulled a bunch of stuff, most of it seemed to go to the main security page at www.postman. Woudln't mind a second pair of eyes here, as this process is new to me.